Add support for different actions
Instead of hardcoding a single action (the execution of a main method),
allow actions to be loaded from the classpath.  Currently only a "run"
action is available.  If "rbt" is run with no arguments the run action is
assumed.  If arguments must be passed in, then applications must be started
as "rbt run [application args]".  For example, examples/netty-client is run
as "rbt run localhost 8080".