Only fetch from remote Git repository when referencing a branch
When a dependency points to a git commit or a git tag don't fetch from the
remote repository if the commit or tag already exists in the cached
repository.

As implemented, repositories are only fetched on every run when following
a raw branch.  If a user enters the a git URI with revision set to
"master~1", the repository will not be fetched on every run, even though
master may have advanced.  This isn't a typical use case, and stopping
unneeded fetches in typical use cases (specifying a tag, commit, or 'git
describe' output) is more important than ensuring that odd revision
specifiers work.