Add URL rewriting and user configuration files
Use URL rewrite rules to allow a user to load artifacts from a local
repository manager (such as Nexus or Artifactory).  Also add default
rewrite rules to enable shorter Maven artifact references.

As part of this change, when a reboot configuration is read the following
files are used (in order of preference):

./reboot-user.xml
~/reboot-user.xml
${reboot-home}/reboot-user.xml
./reboot.xml
${reboot-home}/core/reboot-core.xml

The content from all the files is merged into a single configuration before
the application is started.