Add modules command for printing module fingerprints
The "rbt modules" command prints the current fingerprints of all the
declared modules.  Fingerprints are defined for each type of file system.
The Git filesystem uses the tree ID of the selected revision and path,
while the zip filesystem uses the sha1sum of the zip file.  The standard
file system currently does not support finding fingerprints, so it returns
null.