Edit start script for Mac OS X compatibility
Remove the "readlink -f" calls because "-f" is not supported on all
versions of readlink.  Instead loop over the file until it is no longer
a symbolic link and then use "cd" and "pwd -P" to resolve any remaining
symbolic links in the path.

The readlink call was removed from the build and clean scripts because they
are not expected to be called through symbolic links.